// frontend/src/app/openapi/model/tool-version.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration Manager API
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { ToolVersionConfigurationOutput } from './tool-version-configuration-output';


export interface ToolVersion { 
    /**
     * Unique identifier of the resource.
     */
    id: number;
    name: string;
    config: ToolVersionConfigurationOutput;
}